+++ R/available.R
@@ -73,7 +73,9 @@
 )
 
 .allowedClimateVars_MSY <- c(.allowedClimateVars_M, .allowedClimateVars_S, .allowedClimateVars_Y)
-.allowedClimateVars <- c(.allowedClimateVars_nrm, .allowedClimateVars_MSY)
+## derived here, not by ClimateNA (see calcCumMDC())
+.derivedClimateVars <- c("cumMDC")
+.allowedClimateVars <- c(.allowedClimateVars_nrm, .allowedClimateVars_MSY, .derivedClimateVars)
 
 #' Identify ClimateNA data available via this package
 #'